add doc and interfaces for tmdb services
